Refactor branch navigation and fix PR review comments
- Fixed async/await: Added await to popup.completeAffirmative() calls
- Refactored to use Popup's data-result pattern instead of manual event handlers
  - Removed manual click handlers and event cleanup code
  - Popup now auto-closes and returns selected index
  - Cleaner, more maintainable code following Popup class design
- Fixed CSS: Moved width from :has() selector to .branch-graph-container directly
- Created new branch-graph.js module (613 lines) with well-organized functions:
  - Exported: fetchChatMetadata, traverseBranchTree, updateBranchMetadataAfterRename, showBranchGraph
  - Internal helpers for tree building, layout calculation, and graph rendering
  - Refactored updateBranchMetadataAfterRename into smaller, reusable functions
- Reduced bookmarks.js from ~1250 to ~790 lines (460 lines removed)
- Fixed bug: First branch item (index 0) now navigates correctly
  - Changed result check to distinguish between cancelled (null) and valid index 0
- Fixed tooltip: Changed .mes_branches from data-tooltip to title attribute
  - Now consistent with all other message action buttons
