+++ public/index.html
@@ -7352,7 +7352,7 @@
                                 <div title="Copy" class="mes_button mes_copy fa-solid fa-copy " data-i18n="[title]Copy"></div>
                             </div>
                             <div data-tooltip="Click to open checkpoint chat&#10;Shift+Click to replace the existing checkpoint with a new one" class="mes_button mes_bookmark fa-solid fa-flag" data-i18n="[data-tooltip]Open checkpoint chat&#10;Shift+Click to replace the existing checkpoint with a new one"></div>
-                            <div data-tooltip="Click to view branch chats" class="mes_button mes_branches fa-solid fa-code-branch" data-i18n="[data-tooltip]Click to view branch chats"></div>
+                            <div title="View branch chats" class="mes_button mes_branches fa-solid fa-code-branch" data-i18n="[title]View branch chats"></div>
                             <div title="Edit" class="mes_button mes_edit fa-solid fa-pencil " data-i18n="[title]Edit"></div>
                         </div>
                         <div class="mes_edit_buttons">
