+++ public/scripts/templates/branchList.html
@@ -1,6 +1,6 @@
 <div class="branch-list">
     {{#each branches}}
-    <div class="branch-list-item" data-file-name="" data-branch-index="{{@index}}">
+    <div class="branch-list-item" data-file-name="" data-branch-index="{{@index}}" data-result="{{@index}}">
         <div class="branch-list-item-header">
             <span class="branch-list-item-name"><i class="fa-solid fa-code-branch"></i> <span class="branch-name-text"></span></span>
             <small class="branch-list-item-date">{{formattedDate}}</small>
